refactor: extract default product grid for future default product grid template
